Fixes. fixed quite serious webgl/http bug, infinite read because sockets were blocking (made sockets non-blocking) added REST Manager binary file support Fixed ROOM Tick rate implementation

+++ PiGSFServer/ServerConfig.cs
@@ -17,7 +17,7 @@ namespace PiGSF.Server
         public string avatarUrl = "";
 
         // MMR test
-        public int MMR = 1000;
+        public double MMR = 1000;
 
         public string ToTableString()
         {
